[NFC]: do not implicitly convert the 32bit one constant to a 64bit value (#748)
MODULAR_ORIG_COMMIT_REV_ID: fec533a8431d1c639957718d7b8d58ba9b63444a

+++ GenericML/include/GenericML/Support/TensorEltType.h
@@ -172,7 +172,7 @@ public:
 
   /// Return the width in bits of this type, assuming it is an integer type.
   size_t getIntegerWidthInBits() const {
-    return 1 << getIntegerWidthInLogBits();
+    return size_t(1) << getIntegerWidthInLogBits();
   }
 
   /// Return the width in bits of this type, assuming it is an integer type.
